+++ src/index.js
@@ -1,2 +1,4 @@
 import Composable from './Composable/Composable';
+import sectionifier from './helpers/sectionifier';
 export default Composable;
+export { sectionifier }